requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
psycopg[binary]>=3.2
django-environ>=0.11
requests>=2.32
pytest>=8.0
pytest-django>=4.8
